# personal_financial_app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Finance API
  version: 1.0.0
  description: 'REST API for the Personal Financial App: financial records, bank statement
    PDF extraction, debts, goals, analytics snapshots, exports, and AI-powered analysis,
    categorization and chat (Gemini / OpenAI / Anthropic).'
paths:
...
  /api/currencies/:
    get:
      operationId: currencies_retrieve
      description: Lists the currencies the app can hold amounts in, plus the caller's
        base currency.
      tags:
      - currencies
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/currencies/convert/:
    post:
      operationId: currencies_convert_create
      description: Converts {amount, from, to}. `to` defaults to the caller's base
        currency. Rates come from the daily cache, not a live call.
      tags:
      - currencies
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '503':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT